requirements.txt
flask
python-consul
elasticsearch
loginpass
python-dotenv
python-dateutil
pyyaml
flask-cors
